k8s-upgrades: Pass "do_upgrade" extra var as string
The way we glue together our ansible-playbook command has changed.
The current way the "do_upgrade"-variable is passed leads to a broken
command.
Ansible always has (and still has) problems with Booleans pass as extra
variable. This commit reverts the "pass-the-boolean-as-dict"-workaround
and instead applies the bool-filter everywhere the "do_upgrade"-variable
is used.