State clear dependency on kubectl in stages 4/5
The roles inside of stages 4 and 5 are executed on the same host which
started the run of `ansible-playbook`, i.e., the "orchestrator" (unless,
of course, one uses a custom inventory). Some roles in these stages rely
on `kubectl` for feature which aren't reflected in the ansible modules
yet. We didn't notice so far because the orchestrator usually had
`kubectl` already installed.

This commit improves the documentation by making the dependency on
kubectl explict, by removing duplicates from the list and by removing
the proposal for debian-based systems because the list is not cared for
and might be misleading.