k8s-base: Check all target hosts for host key changes
Before we start the real workload, which may contain sensitive
information, we want to check ALL target hosts for changed host keys.
Previously, we checked only the k8s_nodes here (however, to be fair, the
frontend nodes get checked at a later point in time of the play (again)).