ci: k8s upgrade path: Do not install cert-manager for v1.19
The cert-manager chart requires a Kubernetes version >=1.20.0-0 and
therefore attempting to install it on a v1.19 cluster fails.

As we do not want to bounce back to a cert-manager chart version
supporting v1.19 (because k8s v1.19 is EOL for some time now...), this
commit disables cert-manager for the v1.19 cluster setup. It will be
installed after the cluster has been updated to v1.20 though.