k8s-services: Apply bool filter to enabled switches
A string is interpreted as true by Ansible if it is not empty. That can
lead to unexpected behavior, i.e. if a "SERVICES_enabled" variable is
set to "false" but interpreted as a string, the service gets deployed.
Therefore, we want to apply the bool filter to these variables.