k8s-service-layer: Explicitly add helm repos in advance
We enabled the "update_repo_cache" flag of the Ansible helm module [1]
(and everywhere else helm is used) in !608 to ensure that we always
fetch the lastest changes. However, running "helm repo update" without
a repository being configured results in an error. That is way in !608
prior to the installation of the cert-manager, the corresponding helm
repository is explicitly added. However, it has been missed out that
this needs to be done for all services getting deployed by helm as this
error will occur e.g. for the ingress-nginx service, if cert-manager is
disabled and so on. This commit fixes that.

[1] https://docs.ansible.com/ansible/latest/collections/kubernetes/core/helm_module.html#parameter-update_repo_cache