Add Priority Class to esssential cluster components
This commit adds the priority classes "system-cluster-critical" and
"system-node-critical" to all managed and therefore essential services
and components. These priority classes are already shipped with Kubernetes.
The Kubernetes built-in components itself already have these priority
classes applied. That is also the case for calico.

There is no switch to enable these priority classes, but we do enforce it.
Reasons can be found in the closing paragraph of this commit.
Thankfully, enabling these is very smooth. As most components/services
are either DaemonSets/Deployments and such with a rolling update strategy
or are managed by an operator, Pods get restarted one after another and
the appliance of these priority classes can be considered not disruptive.

Technically, Kubernetes continues working even if one of the managed
services breaks. On an higher level, we want to ensure that these
services and components always work as there are customer and operation
critical. E.g. customer workload may depend on the ingress or cert-
manager being around. From an operation point of view, we want to ensure
that e.g. the etcd-backups are always running.
In a managed cluster, it would be hard to argue that one of the affected
components here has been preempted because it has not been considered
important enough.