CI: Pin ansible-lint image to 5-0.6
There were recent updates to the ansible-lint images [1] which lead to
failures for us. In especially, ansible-galaxy is not around anymore but
we need it to install our collections (devsec hardening ssh [2]). If we
do not install these, the lint job will fail because it can't find the
hardening role. Until a proper fix is implemented, the ansible-lint
image is therefore pinned to the known working image.

[1] https://github.com/cytopia/docker-ansible-lint
[2] https://galaxy.ansible.com/devsec/hardening