Ensure Grafana data can be migrated on upgrades/changes
This does only apply if Grafana uses persistent storage (otherwise we do
not care about its data at all). When changing the Grafana deployment
(e.g. during upgrades and such), the replicas are temporarily scaled to
2. As the new Grafana Pod needs access to the persistent volume and the
access mode of the volume is "ReadWriteOnce" [1], the new Grafana Pod
must be scheduled on the same node as the old one. The simplest way to
ensure that is by using PodAffinity. During a new/fresh installation, if
there is no Pod matching the labelSelector, the PodAffinity gets ignored.

[1] https://kubernetes.io/docs/concepts/storage/persistent-volumes/#access-modes