kubeadm: Explicitly set external cloud provider if on OS
Some fool has accidentally removed this configuration in [1] and that
went unnoticed until now. If we're running the Kubernetes cluster on top
of OpenStack, we need to set "cloud-provider=external" for kubelet, but
not for other Kubernetes components [2][3].

The external cloud provider runs cloud-provider (in our case therefore
OpenStack) specific loops. If it is enabled, information about the
Kubernetes nodes won't be retrieved using local metadata but by the CCM.
This is especially important if scheduling relies on nodes having e.g.
the proper region and AZ tagged.

[1] 6914463add38ecdc1171d422c364c4acd7dcb54d
[2] https://kubernetes.io/docs/tasks/administer-cluster/running-cloud-controller/#running-cloud-controller-manager
[3] https://github.com/kubernetes/cloud-provider-openstack/blob/0418c188d0eb49b7262b7b8f19ccfab318d2cafb/docs/openstack-cloud-controller-manager/using-openstack-cloud-controller-manager.md?plain=1#L36