custom-stage: Change order of inventory flags
The custom stage allows us to trigger roles of all stages. However,
some of the roles of the k8s-base stage depend on the variable storing
the path to the inventory etc directory. This variable is defined in
a vars/-file here [1]. As one can see, it expands the first item of the
variable ansible_inventory_sources. This will lead to a faulty path for
k8s-base roles triggered in the custom-stage which interact with the
etc inventory. To fix this, we can simply change the order in which th
inventories are set for the custom-stage run.

[1] https://gitlab.com/yaook/k8s/-/blob/8b34c4e3dc02d85a6ffbdb987d45d283d7ca3c91/k8s-base/vars/etc.yaml#L1