calico: Allow version mismatch during k8s upgrades
Calico upgrades have been integrated into the k8s-upgrade playbooks.

IPPools should be set up prior to the deployment of calico/node and
Typha on a fresh setup. ÍPPools currently need to be configured via
calicoctl. calicoctl fails if it detects a version mismatch between its
version and the deployed version (which is the version of calico/node
and Typha). That is expected right during the calico upgrade though
because we're updating calicoctl prior to these. We want to keep
configuring the IPPools prior to the deployment/upgrade of calico/node
and Typha though. Therefore, we need to allow the version mismatch if
it is expected.

This commit adds the "--allow-version-mismatch" to the calicoctl call
if we're currently in a Kubernetes upgrade. We're using the command
module to trigger the call. We're passing the command as a list.
We want to set this flag only during upgrades. It will be empty
otherwise. The command (module) will fail if the list contains an empty
element. From this follows, that we need to use the "select" [1] filter
to filter out the empty element before passing the list to the command
module.

The flag is available for calicoctl versions > v3.20. Therefore, the
calico tasks have been removed from the Kubernetes upgrade plays < v1.22.
In these, calicos default version is v3.19 and calico is not getting
updated. Therefore, the tasks would fail because the flag is not available.

[1] https://jinja.palletsprojects.com/en/3.1.x/templates/#jinja-filters.select

f