k8s-upgrades: Overwrite role variables via role parameters
In short: Ansible variable precedences [1] are sometimes... confusing.
Variables passed via "vars:" are scoped to the very specific role only
and not to its dependencies!

This commit changes the Kubernetes upgrade playbooks such that role
variables get overwritten by passing them as role parameters. Passing
them via "vars" scopes the value to the actual role but NOT to it's
dependencies.

The triggering problem: Although we set k8s_version to next_k8s_version
to update calico, that didn't work out and the calico version mapped to
the (previous) kubernetes version was used. This was, because we passed
"k8s_version is next_k8s_version" via "vars:". Because of this, that
change was not passed through to the k8s-config [2] role which is a
dependency of the calico role(s). In that role we basically figure out
which calico version should be deployed. Very confusing and unexpected.
At least for me.

This has only been noticed because the "--allow-mismatch-versions" flag
of calicoctl is only available since version v3.20. calicoctl should be
updated prior to its usage, so that shouldn't be a problem but it was.

[1] https://docs.ansible.com/ansible/latest/user_guide/playbooks_variables.html#understanding-variable-precedence
[2] https://gitlab.com/yaook/k8s/-/tree/devel/k8s-base/roles/k8s-config