linting: Proved linting for yaml syntax
Provide linting for all the y(a)ml files with the help of yamllint within
the gitlab-ci pipeline to check the basic syntax of these files.
Currently there is a long list of warnings and errors, which should be
addressed in the near future. All current errors are (for now) set to the
warning level, to keep the status quo (a green pipeline).

Refer: https://github.com/adrienverge/yamllint