fluxcd: Add a installation process for fluxcd v2
This offers the capability to enable the installation of the flux-system.
Fluxcd is a useful set of controllers that manages the delivery of the cluster
workload in a continuous GitOps manner. Further, this will offer the
opportunity to move the deployment of k8s workload from this repository to a
more dedicated location.

Refer: https://fluxcd.io/