Update Calico to version 3.21
Calico recommends that it's versions should be used with the by them
tested Kubernetes versions.
As we move to Kubernetes 1.22, this would "require" also a newer Calico
version.
Our current Calico templates are based on version 3.19 and are being
templated based on the chosen Kubernetes version.

- templates/config.template.toml
Added a variable, which allows the user to specify one of our supported
calico versions.

- k8s-base/roles/k8s-config/defaults/main.yaml
The calico_version var will reflect the configured Calico version set
within the config.toml. If not specified then the our default
Calico version will be applied.

- k8s-base/roles/calico-control-plane
New calico crds and RBACs for version 3.21.5 have been added.
The RBACs had to be adjusted, as the newer Calico version requires
additional permissions.
The daemonset and deployments have been adjusted according to the
updated Calico typha manifests.
As our current manifests are being used for Calico 3.17 and for 3.19,
but version specific values are also added with each version, we need
to split our templates to support these version specific changes.

- k8s-base/roles/calico
Calico changed the way, how their binaries are being made available
Instead of single downloadable binaries it is now required to
download a whole release-archive. This started with Calico 3.21.3 and
requires us to support both ways.
As the binary archive is very large, the downloaded archive will
be stored for a subsequent run.

- k8s-base/roles/cleanup
A cleanup role has been added to remove the downloaded/extracted
Calico binaries from the orchestration node.