monitoring: Make blackbox exporter version configurable
Make the blackbox exporter helm chart version configurable such that
it can be updated. This has the obvious advantage that one can easily
upgrade if security vulnerabilities get discovered in the deployed
version.

The upgrade to a new version requires no additional steps and therefore
it is enough to pass a new version string to the Ansible helm module.

https://github.com/prometheus-community/helm-charts/tree/main/charts/prometheus-blackbox-exporter#upgrading-chart