workload: Add capability to apply bootstrap manifests
This allows to configure a list of files or globs of kubernetes manifest files,
which will be applied to the cluster. With this functionality, the user is able
to set up initial mechanisms like fluxcd (e.g. GitRepositories and
Kustomizations) that will apply further kubernetes workload in a more
kubernetes native way, than applying resource with ansible.