calico: Change metrics port to 9092
The default Felix metrics port 9091 interferes with the default metrics
port of the rook cephfs plugin [1].

We're changing the calico one as for at least rook v1.3 the port is not
adjustable according to the operator template [2] and I don't want to
test if it works nonetheless. Also, as the calico metrics got merged
recently, changing the calico metrics port requires less adaptions for
existing clusters as it probably is not rolled out much, yet.

[1] https://github.com/rook/rook/blob/fdc2ab55dd5045dada2db16cf5c40fa58577ce97/deploy/examples/operator.yaml
[2] https://github.com/rook/rook/releases/tag/v1.3.9