Add .pre-commit-config.yaml
This commit adds a pre-commit-config to be used with the application
pre-commit[1].

The config runs shellcheck, yamllint and flake8 so errors can
be fixed immediately before the CI even sees them.

It additionally checks for residuals of merges and if executable
scripts have shebangs.

Three mutating hooks are added to fix trailing whitespaces, line endings
and file endings. In case they change anything, the changes need to be
added to be reviewed and staged before trying to commit again.

[1] https://pre-commit.com