rook: Adjust options for manual node/device configuration
This commit adjusts how nodes and their to be used devices can be
configured manually (for bare metal clusters). The existing approach did
not work because everything is flattened [1]]. I didn't want to touch
the inventory updater, yet. It will probably be reworked in the near
future along with the configuration handling. But also, I couldn't wrap
my head around taking a toml, parsing it into json and putting that into
yaml. So I redid the template in yaml.

In addition, this commit changes the defaults such that rook can be
deployed on OpenStack (Cinder) without further configuration.

A new variable "use_all_available_nodes" has been introduced to
configure this behavior (described by the name) of rook.

[1] https://gitlab.com/yaook/k8s/-/blob/bb7e95c2c2e6258e9b0d485ca9fdb5e42d5be46b/actions/update_inventory.py#L160-183