Add role for Helm-based installation of rook-ceph
Starting with rook v1.7, an official Helm chart is provided and has become the
recommended installation method. The charts take care most installation and
upgrade processes. This commit adds a new role rook_v2 which includes support
for the Helm-based installation as well as a migration path from rook_v1.