rook: Rework version handling
Unfortunately, !674 didn't fix the actual issue. A ceph upgrade failed
because the string cutting didn't work out. The problem was that the
"rook_conf_major_version" and its comrades had no fitting values when a
Ceph upgrade gets triggered.

In addition, the "wait_for_{rook,ceph}_upgrade.yaml" [1] tasks failed with:
ERROR! Unexpected Exception, this is probably a bug: unsupported operand
type(s) for |=: 'bool' and 'AnsibleUnsafeText'

The version handling in the rook role is really complex and confusing.
This commit attempts to simplify it a bit by moving the version splitting
of the rook version into its major and minor version parts to the defaults
of the role. The version checks [2] (which are used to figure out if
(a valid) upgrade is attempted) have been simplified along.

As the Ceph version is not individually configurable but mapped to the
rook version, we don't need to version check it so fastidiously.
The upgrade files have been renamed so we can include them more easily [3].

[1] https://gitlab.com/yaook/k8s/-/blob/5d7b1df339fb23669e4bc3d7f7e58104b51366aa/k8s-service-layer/roles/rook_v1/tasks/wait_for_ceph_upgrade.yaml
    https://gitlab.com/yaook/k8s/-/blob/5d7b1df339fb23669e4bc3d7f7e58104b51366aa/k8s-service-layer/roles/rook_v1/tasks/wait_for_rook_upgrade.yaml
[2] https://gitlab.com/yaook/k8s/-/blob/5d7b1df339fb23669e4bc3d7f7e58104b51366aa/k8s-service-layer/roles/rook_v1/tasks/version_checks.yaml
[3] https://gitlab.com/yaook/k8s/-/blob/5d7b1df339fb23669e4bc3d7f7e58104b51366aa/k8s-service-layer/roles/rook_v1/tasks/version_checks.yaml#L104-109