k8s-worker: Let join attempt timeout
By default the shell module has no timeout. We especially in the CI
observed that a worker is not able to join the cluster for whatever
reasons, but the task neither fails not timeouts until the hard
Gitlab CI job timeout of 6 hours hits.

With the async and delay options we define the maximum number of
seconds the task is allowed to run and the polling interval. This will
cause the task to timeout.