Retry kubernetes.core.k8s tasks
During the usage of the kubernetes.core.k8s module different errors can
happen. All of these got observed in the CI and led to failures.

If one has bad luck, the etcd leader changes and the task returns an
internal error (500). It is also possible that the CRD of a custom
resource can't be validated in time, but we do set validate.fail_on_error
to true for nearly all tasks. Other than that, the usual internet
storminess may hits.

This commit adds retries to all kubernetes.core.k8s tasks so that we
don't fail hardly in such occasions.