system-update: Set KUBECONFIG in action script
In between the system updates (in the respective playbook) we want to
check that the rook/ceph cluster is healthy prior to the drain of the next
node (they get drained in serial). However, the health check did not
work because no kubeconfig has been set and in addition to that, we did
ignore that. That led to all the health check tasks being ignored :O

This hasn't been noticed, because when setting up a toycluster one
in almost all cases sets the env variable KUBECONFIG on its own and then
the task is able to succeed.

To be sure that KUBECONFIG is set, we export it in the system update
node action script.