k8s-worker: Join cluster with slight delay
We're sometimes observing problems with worker join attempts in the CI.
The root cause is unclear, yet. It looks like random network failures.

Letting the workers join with a slight delay may help here. The join
attempts are triggered simultaneously due to the nature of Ansible.
Although we never tested that, that probably does not scale well.
A slight random delay ensures that the k8s CP is not flooded at once.