rook: Bump default version to v1.7.11
The config.toml template already has been adjusted in [1]. v1.2.3 is
too old and missing too much features to be the default anymore.
Existing clusters not explicitly configuring the rook version will
notice that the default version has been bumped by observing that the
rook_v1 role will fail on a subsequent run as rook has to be updated
from minor to minor version (so 1.2.3 -> 1.7.11 is not possible at once).
The rook_v1 role catches that (and also if an update is configured but
disruption is not allowed) and will fail. Explicitly pinning the rook
version to v1.2.3 and then update these clusters step by step is a
smaller burden than leaving the default to v1.2.3 and create new clusters
with a granny rook version.

[1] b19ff0c33a76301b9c27f028afa1b0d4afc99f89