rook: Enable monitoring
This commit enabled Prometheus monitoring for rook [1] (if the
Prometheus stack is enabled). Instead of deploying an additional
Prometheus instance as it is supposed in [1], we're using the existing
one. The ServiceMonitor for the rook ceph mgr already is implemented
in [2] and also tested in [3]. Enabling metrics requires the respective
RBAC resources though, otherwise the rook operator won't be fully
functional (i.e. when creating a new cluster it refuses to create the
Ceph OSDs because it fails creating the ServiceMonitor).

It is important to set the respective monitoring labels if needed.

[1] https://rook.io/docs/rook/v1.7/ceph-monitoring.html
[2] https://gitlab.com/yaook/k8s/-/blob/e4a267d65ec994e783ebd4a1f97067b428db113d/k8s-managed-services/roles/monitoring_v2/tasks/main.yaml#L79-92
[3] https://gitlab.com/yaook/k8s/-/blob/e4a267d65ec994e783ebd4a1f97067b428db113d/k8s-managed-services/test-roles/monitoring_v2_test/tasks/scrape_managed_services.yaml#L31-44