nginx-ingress: Add monitoring
This commit conditionally enabled Prometheus monitoring of the nginx-
ingress controller. To properly parse the common service monitor labels,
we need jinja2. For easier comprehension, the helm values of the nginx-
ingress chart are moved to a template file.

In addition, a respective Grafana dashboard [1] is deployed if Grafana
is enabled.

[1] https://grafana.com/grafana/dashboards/9614-nginx-ingress-controller/