etcd-backup: Don't skip helm deployment on first run
... if the bucket does not exist, yet. The previous tasks ensure that
the bucket gets created and the etcd-backup chart should be deployed
in anyway. The "when:" condition is a leftover of a redo in !560 and
shouldn't be here anymore. With the condition, a bucket gets created,
but the etcd-backup does not get deployed on the first run (but a second
execution is needed).