Implement docker-to-containerd (CRI) migration path
This commit implements a playbook allowing to migrate all nodes of a
cluster from using to docker as CRI to the usage of containerd.
A wrapper (bash) script has been added so that the playbook can easily
be triggered. The necessary steps have been documented and the changelog
file has been updated accordingly.

After a successful migration, one must adjust its config/config.toml.

The implementation of this migration path is a prerequisite to the
implementation of Kubernetes v1.24 support.