CI: Change ansible-lint image
From now on we're using a python image and are installing ansible, our
requirements and ansible-lint into it. The reason for that is, that the
ansible-lint image previously used is updated inregularly and we want to
ensure the linting is in sync with the ansible version used.

For now, we do ignore all failures coming up with that new procedure.
These will be fixed in followup commits.