Ensure tests are executed, redo delegation
This commit fixes the following bug: The checks if the service
(un)reachable from the target node have been skipped because we targeted
master[0] but added this condition [1] that the tasks get skipped if so.

With this commit we're now targeting the last k8s_node (which in a single
node setup is master[0] and then the when condition hits and that is
wanted). This required to delegate all the tasks which talk to the
Kubernetes API to localhost (or to a master node, but that would make
things even more complicated).

To ensure the echo deployment Pods do not get scheduled on the last
k8s_node by chance (which is targeted and which could then reach the Pod
if it is scheduled on it and therefore the tasks would make no sense),
node anti affinity has been added.

[1] https://gitlab.com/yaook/k8s/-/blob/b11be4e8a04de5d841de21834722ed817f3c53dc/k8s-base/test-roles/check-networkpolicy-enforcement/tasks/main.yaml#L137