containerd: Move to package from docker registry
We saw issues in the CI and with local dev-clusters when creating new
clusters. kubeadm couldn't download the Kubernetes images and a run
failed at [1]. Apparently, there are problems with the containerd
provided by Ubuntu.

Therefore, we're switching to a newer containerd version from the docker
repository. That package is called "containerd.io" and mentioned in the
official containerd installation docs anyway [3].

[1] https://gitlab.com/yaook/k8s/-/blob/90e52a53ec329d82a6c2dd18a976f4cd804dc456/k8s-base/roles/k8s-master/tasks/main.yaml#L92
[2] https://forum.linuxfoundation.org/discussion/862825/kubeadm-init-error-cri-v1-runtime-api-is-not-implemented
[3] https://github.com/containerd/containerd/blob/main/docs/getting-started.md#option-2-from-apt-get-or-dnf