containerd: Restart systemd service after node setup
It is not necessary to restart kubelet and drain the nodes one after
another to geth them into the ready state. That procedure was really
time consuming :/.
It is sufficient to restart the containerd systemd service once after
setting up the node and its CNI. Why that is needed continues to remain
unclear for now, however.