ansible-lint: Fix fqcn[keyword], avoid collections
Although we're using the full fqcn for the devsec ssh hardening role,
the ansible-lint complains that the collections specification should be
avoided:
fqcn[keyword]: Avoid `collections` keyword by using FQCN for all plugins,
modules, roles and playbooks.

Therefore, this commit removes it but adds a comment to make clear where
the role it comes from.