Add failed_when condition
We observed confusion with the command module and
"2> /dev/null" (it sometimes in rare occasions complains that the
"/dev/null" module cannot be found(at least that happened more than one
in the CI)). To prevent this, we're allowing the return codes 0 and 1
only now instead. The command module has no diff mode
available anyway and therefore won't spam the output even if we're not
redirecting.