Embed GPG keyrings in corresponding ansible roles
While this brings additional maintenance effort (we'll have to update
the keyring if it ever changes), it has the advantage that we do not
need to access the internet for downloading the keyring.

"But we'll have to access the internet to download the packages!", you
might observe. For that use case, we have an apt-proxy in such setups,
under development in a different merge request.