vault: include, not import backup tasks
include_* and import_* are strange beasts in ansible. While tasks/roles
which are "included" are loaded when the play hits the include_*
directive, stuff is already pre-parsed when it's import_*-ed. For some
reason, the logic in backup.yaml wasn't guarded enough (we're setting
certain variables during runtime and looping over them), so the play
fails if backup=false (because vars inside that task which shouldn't be
considered in first place!).