Implement workflow to pivot a cluster to use its own Vault
This required some minor changes to the vault_v1 role. In particular, we
need to allow NodePort type services.

These were needed anyway for `management_cluster_integration`, even
though that was apparently not implemented. In addition, we now add the
IPs of all nodes to the certificate based on whether the `service_type`
is `NodePort` (or "higher"), and not based on whether
`management_cluster_integration` is enabled (however, we default
`service_type` to `NodePort` if `management_cluster_integration` is
set).

All the rest should be explained in the extended documentation.