Export KUBECONFIG for stage3
This allows us to interact with the Kubernetes API (after it has been
spawned) from the orchestration (local-) host. This is handy and wanted,
because we can then avoid the delegation to master[0] for such tasks.
master[0] may not be available which would lead to failures of these tasks.

We should find all occasions of tasks which interact with the Kubernetes
API and are delegated to master[0] and switch the delegation to
localhost instead.