calico: Add comprehensive migration tasks
This commit adds step to migrate from an manifest-based (Calico the hard
way) installation to the Tigera operator (operator-based installation).

Unfortunately, we customized our Calico installation to such an extend,
that the automated operator migration is not possible
as the Tigera operator is unable to adopt the existing resources.

Therefore, we're removing EVERYTHING that is calico-related and set up
via the manifest-based installation. This includes binaries on the nodes,
monitoring resources, files in the cluster repository inventory and more.
That is needed so that the Tigera operator can start on a green field
and we're not going to run into regressions (we're actually life-cycling
our clusters and not just delete and recreate everything, y'know).

That the previous approach was not clean could already be seen in the
necessary Tigera operator restart and node label patch workaround.