Tigera: Rework setup procedure
Add support for existing calico customization options. These options can
all be enabled/configured via a few lines in the Installation manifest.
There is no reason we should not do that (and a lot of why we must).

Instead of throwing the tasks against master-0, talk to the Kubernetes
API from the orchestration node.
Also add retries to these tasks to be more resilient to network hiccups.

Although we're creating the calico API servers, we still install
calicoctl on the CP nodes to be able to tinker with calico custom
resources even if the API server pods are down.

As we're including a tasks of the calico-control-plane role and that
role has a dependency on the "calico" role, we need to guard the tasks
of the "calico" role.

Make the calicoctl binary source configurable (!703)