calico: Switch to Tigera operator helm chart
As the URL of the Tigera operator changes among calico versions, the
previous approach to download the operator manifest and apply it
manually would not have been trivially to maintain for past and furture
versions.
Instead, we're now using the helm chart to deploy the Tigera operator
which save us from such burdens.
It's still possible to pass a custom values file, so we're keeping
wanted flexibility here.