terraform: Allow separate volume type specification
Previously, it was not possible to specify a differing volume type for
root/boot disk volumes for each instance. However, one may want to have
differing volume types for e.g. masters and workers.

This commit allows to specify volume types for each instance separately.
I.e. for masters and workers. The default volume type to be used can
still be set via "root_disk_volume_type". Otherwise the default of the
IaaS environment is used.

For now, the volume type for gateways can't be specified individually.
One reason for that is, that one can't specify the number of gateways
either as there is one gateway per availability zone. Also we should
keep things simple if possible.