calico: Don't expose tigera operator switch
We decided to enforce the operator starting with k8s v1.25, leave it
optional for k8s v1.24, but don't make it freely decidable for future
versions. Therefore, we don't expose the variable allowing to switch
the operator on/off that obviously.

Adding support for Kubernetes versions below v1.24 requires additional
development overhead, as the current procedure does not work for the
mapped Tigera calico helm releases. We want to avoid that as all these
Kubernetes versions are EOL anyway. This has been addressed and
explained in another commit.

We're moving the respective variable to the "k8s-config" role so that we
can easily enable the Tigera operator by default for the upcomming
K8s v1.25 support. As we want to avoid in-line defaults, especially
if they are needed in multiple places, we need to circumvent that
the tasks defined in the "calico" role get executed if the Tigera
operator is configured by adding a meta end_play task for that case.

Also, there is no need to check if the CNI is configured to be calico,
as that's the only supported one now.

To keep things simple and speed development up, the change is put
into this separate commit instead of merging the changes back into the
existing commits.