Drop unnecessary group creation
The "god" group has special meaning in the C&H world. It should not be
included in yaook/k8s. Also there is no gain in giving members of this
group access to the kubeconfig. All of these can escalate to root
anyway. Therefore, we're dropping these tasks here.

That requires to use "become: true" for all tasks which have masters as
target and use the kubeconfig; or delegate them to localhost.