kubelet: Set eviction hard thresholds
To prevent system OOMs, we're enforcing an eviction (hard) threshold
of 256Mi from now on. The default is 100Mi.
This should be sufficient to prevent the system from system OOMing before
kubelet is able to evict Pods. We noticed that the default value is not
sufficient to react in time under certain circumstances.

The other values are the defaults for Linux nodes, but one has to define
all if one custom value is used, otherwise the not specified ones are set
to null [2].

There is good documentation about eviction [1].

We're also specifying a soft eviction threshold of 384Mi to allow
gracefully shutting down Pods instead of directly hammer slamming them.
There are no defaults for soft eviction.
One must define the evictionSoftGracePeriod for a customized soft
eviction threshold.

[1] https://kubernetes.io/docs/concepts/scheduling-eviction/node-pressure-eviction
[2] https://kubernetes.io/docs/concepts/scheduling-eviction/node-pressure-eviction/#hard-eviction-thresholds