refactor: rename on_openstack ansible variable
For the "core split", we want to cut out the part that does specific things for
a kubernetes cluster on openstack. The first step is to refactor the
"on_openstack" (boolean) variable, which appears in some places in the code,
with the more general "infrastructure_type" (string) variable.

This variable is set to "infrastructure_type=openstack" by terraform just like
it was done with "on_openstack = true". This rotates the handover/understanding
of this parameter in the right direction and couples it more loosely. It also
offers the possibility of adding other infrastructure types. The related logic
in the ansible playbooks has been adjusted.

See: https://gitlab.com/yaook/k8s/-/issues/556