Allow PGs to be in any "active+clean" state
It is possible for health checks to run while deep scrubbing is taking
place, e.g. if a PG enters deep scrub right after having recovered or if
this is the first node which is being drained and there was a deep scrub
scheduled for just now.

As the goal of the health check is to ensure that taking down any OSD
node would not break anything, any state which contains active+clean is
okay, so we accept any state instead of strictly accepting only
active+clean.