Add https to service-account-issuer apiserver cmd
Fixes #559

This is necessary to enable the ServiceAccountIssueDiscovery, which is
required for all k8s-conformance tests to pass.

The --service-account-issuer command line argument for kube-apiserver
tells the apiserver which issuer should be put into the service account
tokens.

For federation to work correctly, the issuer should be an OIDC conformant
HTTPS URL and should have issuer information available at .well-known/openid-configuration.
Kube-apiserver only exposes this endpoint if --service-account-issuer is
an HTTPS URL.