Apply changed apiserver arg to exist. deployments
References #559

This commit patches the kubeadm-config ConfigMap when an existing
installation is being upgraded to assure that the --service-account-issuer
extra argument to kube-apiserver is set correctly.

It is assumed that all clusters with a k8s version >= 1.25 already
have this change applied. So there exists a condition that
skips this check for k8s target versions >= 1.26.