Drop support for Kubernetes v1.19
This version is EOL for quite some time and there are no clusters in the
wild using that version anymore (at least none we're aware of).
This step has been announced several times in the Shore Leave.

Unfortunately, for the CI upgrade job we still need the workaround as
the services require v1.21 as minimum in the mean time..