terraform: add config_drive to ignore list
As the config drive is not used by yaook/k8s, changing this value should
not trigger re-creation of an instance. This commit adds config_drive
to the ignore lists of all instances, which can be useful when
importing an existing instance that has been created or modified outside
of Terraform.