Add kubeadm patch to add kube-apiserver arg
Because the `--service-account-issuer` argument of kube-apiserver
has been changed from `kubernetes.default.svc` to
`https://kubernetes.default.svc`, problems arise because
kube-apiserver does not see old tokens (with old issuer included)
as valid.

This is fixed by adding the `--service-account-issuer` argument
two times. The first time for the issuer that is used for new
tokens and a second time for the apiserver to still see tokens
with the old issuer value as valid.

Due to the fact that `kubeadm-config` does not allow to specify
additional apiserver arguments multiple times
(https://github.com/kubernetes/kubeadm/issues/1601),
it was required to create a patch file which is then used when
executing `kubeadmin upgrade {apply,node}`.