Move NVIDIA DCGM Exporter to KSL
Supporting machines with NVIDIA GPUs is not an advanced, but basic
feature. The DCGM exporter is part of our monitoring stack and therefore,
we should deploy the helm chart in the KSL layer. The KMS layer is going
to be removed anyway in the near future during the so called "kore split".
(join the shore leave for more information)

From now on we're also always deploying the latest version by default as
we are not able to keep pace with version pinning.