Fix LBaaS priority inconsistency
Until now, the vrrp_priorities set in config/config.toml where used for
the VRRP VIP in roles/keepalived, but not for roles/ch-k8s-lbaas-agent,
as it wasn't available in the stage. So the lbaas always used the
default setting in roles/vrrp-common which led to inconsitencies
especially when the set vrrp_priorities not have been in descending
order, as the default values are. That problem is fixed now.

Fixes #547.